Support drawing to multiple framebuffers with different scales
Provide an API to configure an outputs grid for a stage. Given this
grid, clutter will draw to each given rectangle using a separate
framebuffer. Each such framebuffer may have a different scale.

This patch implements this only for the X11 backend, and it currently
only supports drawing each output to an offscreen framebuffer which is
then composited on the onscreen framebuffer for each frame. This is
to simplify input event handling as it currently depends on only having
one X11 window. The extra composite could be avoided for single output
situations (running as X11 CM) without having to deal with multiple X11
windows.

For the native (KMS) backend, each output should simply be one onscreen
framebuffer, and thus there will be no extra composite, and as a bonus,
potentially reduced memory usage.

This patch currently breaks every backend except the X11 backend, so
those have to be disabled to be able to build clutter. This is because
the ClutterStageCogl does no longer have a single onscreen framebuffer
to depend on.

This patch currently also breaks drawing without setting the output grid.